first revision of the formatting mechanism
Thanks to u/xe3to on Reddit for the advices and inspiration.
Now the program outputs constellations info from non-preformatted files: it fills a template file with the data from constellation files, formats it and prints it out.
This way it's easier to store constellation data, and overall more polished.